pyproject.toml
[project]
name = "web-crawler-mcp-server"
version = "1.0.0"
description = "MCP server for web crawling with newspaper4k integration"
authors = [
    {name = "Sai Min Htet", email = "saiminhtet@u.nus.edu"},
]
dependencies = [
    "mcp>=1.0.0",
    "newspaper4k>=0.2.8",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "playwright>=1.40.0",
    "python-dateutil>=2.8.0",
    "lxml>=4.9.0",
    "feedparser>=6.0.0",
]

[project.scripts]
web-crawler-mcp-server = "src.web_crawler_mcp.server:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"